feat: change sdk action
