+++ frontend/src/component/onboarding/SelectSdk.tsx
@@ -1,23 +1,8 @@
 import { Avatar, Box, Link, styled, Typography } from '@mui/material';
 import type { FC } from 'react';
-import android from 'assets/icons/sdks/Logo-android.svg';
-import dotnet from 'assets/icons/sdks/Logo-net.svg';
-import flutter from 'assets/icons/sdks/Logo-flutter.svg';
-import go from 'assets/icons/sdks/Logo-go.svg';
-import swift from 'assets/icons/sdks/Logo-swift.svg';
-import java from 'assets/icons/sdks/Logo-java.svg';
-import javascript from 'assets/icons/sdks/Logo-javascript.svg';
-import node from 'assets/icons/sdks/Logo-node.svg';
-import php from 'assets/icons/sdks/Logo-php.svg';
-import python from 'assets/icons/sdks/Logo-python.svg';
-import react from 'assets/icons/sdks/Logo-react.svg';
-import ruby from 'assets/icons/sdks/Logo-ruby.svg';
-import rust from 'assets/icons/sdks/Logo-rust.svg';
-import svelte from 'assets/icons/sdks/Logo-svelte.svg';
-import vue from 'assets/icons/sdks/Logo-vue.svg';
 import { formatAssetPath } from 'utils/formatPath';
 import { SectionHeader } from './SharedComponents';
-import type { ClientSdkName, Sdk, ServerSdkName } from './sharedTypes';
+import { clientSdks, type Sdk, serverSdks } from './sharedTypes';
 
 const SpacedContainer = styled('div')(({ theme }) => ({
     padding: theme.spacing(5, 8, 8, 8),
@@ -66,27 +51,6 @@ const StyledAvatar = styled(Avatar)(({ theme }) => ({
     left: theme.spacing(2),
     boxShadow: theme.shadows[2],
 }));
-
-export const serverSdks: { name: ServerSdkName; icon: string }[] = [
-    { name: 'Node', icon: node },
-    { name: 'Golang', icon: go },
-    { name: 'Ruby', icon: ruby },
-    { name: 'PHP', icon: php },
-    { name: 'Rust', icon: rust },
-    { name: 'DotNet', icon: dotnet },
-    { name: 'Java', icon: java },
-    { name: 'Python', icon: python },
-];
-
-export const clientSdks: { name: ClientSdkName; icon: string }[] = [
-    { name: 'Javascript', icon: javascript },
-    { name: 'React', icon: react },
-    { name: 'Vue', icon: vue },
-    { name: 'Svelte', icon: svelte },
-    { name: 'Swift', icon: swift },
-    { name: 'Android', icon: android },
-    { name: 'Flutter', icon: flutter },
-];
 
 interface ISelectSdkProps {
     onSelect: (sdk: Sdk) => void;
